to BigDecimal

+++ src/main/java/app/holdingUnits/Edge.java
@@ -1,5 +1,7 @@
 package app.holdingUnits;
 
+import java.math.BigDecimal;
+
 /**
  * Edge which connects two nodes and contains converting quotient.
  *
@@ -16,7 +18,7 @@ public class Edge {
     public Node node2;
 
     /** the quotient of converting. */
-    public Double quotient;
+    public BigDecimal quotient;
 
     /**
      * constructs edge.
@@ -24,7 +26,7 @@ public class Edge {
      * @param node2 node to where converting is going.
      * @param quotient the quotient of converting.
      */
-    public Edge(Node node1, Node node2, Double quotient) {
+    public Edge(Node node1, Node node2, BigDecimal quotient) {
         this.node1 = node1;
         this.node2 = node2;
         this.quotient = quotient;
@@ -42,7 +44,7 @@ public class Edge {
      * gets the quotient of converting.
      * @return the quotient.
      */
-    public Double getQuotient() {
+    public BigDecimal getQuotient() {
         return quotient;
     }
 }
